+++ lib/l10n/app_localizations_ru.dart
@@ -1024,6 +1024,9 @@ class AppLocalizationsRu extends AppLocalizations {
   String get episodes => 'Эпизоды';
 
   @override
+  String get next_episode => 'Следующий эпизод';
+
+  @override
   String get categories => 'Категории';
 
   @override